LayoutTests/http/tests/inspector/target/provisional-load-cancels-previous-load.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../resources/inspector-test.js"></script>
<script>
function test()
{
    let lastAssignedId = 0;
    let targetToId = new Map;
    function stableTargetId(targetId)
    {
        let id = targetToId.get(targetId);
        if (!id) {
            id = ++lastAssignedId;
            targetToId.set(targetId, id);
        }
        return id;
    }

    let suite = InspectorTest.createAsyncSuite("Target.PSON");

    suite.addTestCase({
        name: "ProvisionalPagePaused",
        description: "Check that new provisional page can be paused before navigation.",
        test(resolve, reject) {
            InspectorTest.log(`Current target is ${stableTargetId(WI.mainTarget.identifier)}.`);
            const url = "http://localhost:8000/inspector/target/provisional-load-cancels-previous-load.html";
            let navigatedTwice = false;

            WI.targetManager.addEventListener(WI.TargetManager.Event.TargetAdded, (event) => {
                let target = event.data.target;
                let targetId = stableTargetId(target.identifier);
                InspectorTest.pass(`Should receive TargetAdded event for target ${targetId}.`);
                InspectorTest.expectTrue(target.isProvisional, `Target ${targetId} should be provisional.`);
                InspectorTest.expectTrue(target.isPaused, `Target ${targetId} should be paused on start.`);
                if (!navigatedTwice) {
                    navigatedTwice = true;

                    // Send two consequtive navigation requests. The latter will cancel provisional
                    // load of the former.
                    WI.mainTarget.PageAgent.navigate(url);
                }
            });

            WI.targetManager.addEventListener(WI.TargetManager.Event.DidCommitProvisionalTarget, (event) => {
                let {previousTargetId, target} = event.data;
                InspectorTest.pass(`Should receive DidCommitProvisionalTarget event ${stableTargetId(previousTargetId)} => ${stableTargetId(target.identifier)}.`);
            });

            WI.targetManager.addEventListener(WI.TargetManager.Event.TargetRemoved, (event) =>{
                let targetId = stableTargetId(event.data.target.identifier);
                InspectorTest.pass(`Should receive TargetRemoved event for target ${targetId}`);
            });

            // Wait for page reload event to avoid race between test results flushing and the test completion.
            InspectorTest.awaitEvent(FrontendTestHarness.Event.TestPageDidLoad)
            .then(() => {
                InspectorTest.expectEqual(lastAssignedId, 3, `Should have seen 3 different targets.`);
            })
            .then(resolve);

            WI.mainTarget.PageAgent.navigate(url);
        }
    });

    suite.runTestCasesAndFinish();
}
</script>
</head>
<body onload="runTest()">
<p>Test that two consequtive cross domain navigation requests will result in two
provisional targets being created, first of which is later destroyed and the second
is committed.</p>
</body>
</html>
