LayoutTests/http/tests/inspector/target/provisional-load-cancels-previous-load-expected.txt

Test that two consequtive cross domain navigation requests will result in two provisional targets being created, first of which is later destroyed and the second is committed.


== Running test suite: Target.PSON
-- Running test case: ProvisionalPagePaused
Current target is 0.
PASS: Should receive TargetAdded event for target 1.
PASS: Target 1 should be provisional.
PASS: Target 1 should be paused on start.
PASS: Should receive TargetRemoved event for target 1
PASS: Should receive TargetAdded event for target 2.
PASS: Target 2 should be provisional.
PASS: Target 2 should be paused on start.
PASS: Should receive TargetRemoved event for target 0
PASS: Should receive DidCommitProvisionalTarget event 0 => 2.
PASS: Should have seen 3 different targets.
